+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 - Class ``TechnicalVideoMetadata`` for creating technical video metadata in VideoMD standard
 - Class ``TechnicalCSVMetadata`` for creating technical CSV metadata in ADDML standard
 - Classes ``TechnicalFileObjectMetadata`` and ``TechnicalBitstreamObjectMetadata`` for representing file and bitstream technical metadata respectively using PREMIS objects
+- Identical technical media metadata entries will automatically be deduplicated in the generated METS if possible
 
 0.1.0 - 2023-06-14
 ------------------